Move CloudManager to the profile module
